Restricting access further.

+++ source/ch/dbs/actions/maintenance/BulkOperations.java
@@ -38,64 +38,69 @@ import enums.Result;
  * List the possible options for the maintenance of a given account.
  */
 public class BulkOperations extends DispatchAction {
-
+    
     public ActionForward deleteorders(final ActionMapping mp, final ActionForm form, final HttpServletRequest rq,
             final HttpServletResponse rp) {
-
-        String forward = Result.FAILURE.getValue();
+        
         final Auth auth = new Auth();
-
+        // if activated on system level, access will be restricted to paid only
+        if (auth.isPaidOnly(rq)) {
+            return mp.findForward(Result.ERROR_PAID_ONLY.getValue());
+        }
+        
+        String forward = Result.FAILURE.getValue();
+        
         // catching session timeouts
         if (auth.isLogin(rq)) {
             // access restricted to librarians and admins only
             if (auth.isBibliothekar(rq) || auth.isAdmin(rq)) {
-
+                
                 final MaintenanceForm operation = (MaintenanceForm) form;
                 final UserInfo ui = (UserInfo) rq.getSession().getAttribute("userinfo");
                 final Text cn = new Text();
-
+                
                 try {
-
+                    
                     if (operation.isConfirmed()) {
                         // execute operation
                         final int deleted = operation.deleteOrders(ui, cn.getConnection());
-
+                        
                         // set result from DELETE back into form
                         operation.setNumerOfRecords(deleted);
-
+                        
                         final Message msg = new Message();
                         msg.setMessage("maintenance.numberOfRecordsDeleted");
                         msg.setSystemMessage(String.valueOf(operation.getNumerOfRecords()));
                         msg.setLink("maintenance.do");
                         rq.setAttribute("message", msg);
-
+                        
                         forward = Result.SUCCESS.getValue();
                     } else {
                         // force confirmation of operation
                         forward = "confirm";
-
+                        
                         // get number of records involved
                         operation.setNumerOfRecords(operation.countDeleteOrders(ui, cn.getConnection()));
-
+                        
                         operation.setMethod("deleteorders");
                         rq.setAttribute("operation", operation);
                     }
-
+                    
                 } finally {
                     cn.close();
                 }
-
+                
                 // navigation: set 'account/konto' tab as active
                 final ActiveMenusForm mf = new ActiveMenusForm();
                 mf.setActivemenu("konto");
                 rq.setAttribute(Result.ACTIVEMENUS.getValue(), mf);
-
+                
             } else {
                 final ErrorMessage m = new ErrorMessage("error.berechtigung");
                 m.setLink("searchfree.do");
                 rq.setAttribute(Result.ERRORMESSAGE.getValue(), m);
             }
-
+            
         } else {
             final ActiveMenusForm mf = new ActiveMenusForm();
             mf.setActivemenu(Result.LOGIN.getValue());
@@ -103,67 +108,72 @@ public class BulkOperations extends DispatchAction {
             final ErrorMessage em = new ErrorMessage("error.timeout", "login.do");
             rq.setAttribute(Result.ERRORMESSAGE.getValue(), em);
         }
-
+        
         return mp.findForward(forward);
     }
-
+    
     public ActionForward deleteusernoroders(final ActionMapping mp, final ActionForm form, final HttpServletRequest rq,
             final HttpServletResponse rp) {
-
-        String forward = Result.FAILURE.getValue();
+        
         final Auth auth = new Auth();
-
+        // if activated on system level, access will be restricted to paid only
+        if (auth.isPaidOnly(rq)) {
+            return mp.findForward(Result.ERROR_PAID_ONLY.getValue());
+        }
+        
+        String forward = Result.FAILURE.getValue();
+        
         // catching session timeouts
         if (auth.isLogin(rq)) {
             // access restricted to librarians and admins only
             if (auth.isBibliothekar(rq) || auth.isAdmin(rq)) {
-
+                
                 final MaintenanceForm operation = (MaintenanceForm) form;
                 final UserInfo ui = (UserInfo) rq.getSession().getAttribute("userinfo");
                 final Text cn = new Text();
-
+                
                 try {
-
+                    
                     if (operation.isConfirmed()) {
                         // execute operation
                         final int deleted = operation.deleteUserNoOrders(ui, cn.getConnection());
-
+                        
                         // set result from DELETE back into form
                         operation.setNumerOfRecords(deleted);
-
+                        
                         final Message msg = new Message();
                         msg.setMessage("maintenance.numberOfRecordsDeleted");
                         msg.setSystemMessage(String.valueOf(operation.getNumerOfRecords()));
                         msg.setLink("maintenance.do");
                         rq.setAttribute("message", msg);
-
+                        
                         forward = Result.SUCCESS.getValue();
                     } else {
                         // force confirmation of operation
                         forward = "confirm";
-
+                        
                         // get number of records involved
                         operation.setNumerOfRecords(operation.countDeleteUserNoOrders(ui, cn.getConnection()));
-
+                        
                         operation.setMethod("deleteusernoroders");
                         rq.setAttribute("operation", operation);
                     }
-
+                    
                 } finally {
                     cn.close();
                 }
-
+                
                 // navigation: set 'account/konto' tab as active
                 final ActiveMenusForm mf = new ActiveMenusForm();
                 mf.setActivemenu("konto");
                 rq.setAttribute(Result.ACTIVEMENUS.getValue(), mf);
-
+                
             } else {
                 final ErrorMessage m = new ErrorMessage("error.berechtigung");
                 m.setLink("searchfree.do");
                 rq.setAttribute(Result.ERRORMESSAGE.getValue(), m);
             }
-
+            
         } else {
             final ActiveMenusForm mf = new ActiveMenusForm();
             mf.setActivemenu(Result.LOGIN.getValue());
@@ -171,8 +181,8 @@ public class BulkOperations extends DispatchAction {
             final ErrorMessage em = new ErrorMessage("error.timeout", "login.do");
             rq.setAttribute(Result.ERRORMESSAGE.getValue(), em);
         }
-
+        
         return mp.findForward(forward);
     }
-
+    
 }
